Reorganize and document
 - split modules inot utility, testdata, diredit
 - use test_testdata as  a usage example
 - reflest testdata usage example in usage docs
 - update build and contrib docs
