+++ .bumpversion.cfg
@@ -3,8 +3,6 @@ current_version = 0.1.2
 commit = True
 tag = True
 
-[bumpversion:file:HISTORY.rst]
-
 [bumpversion:file:setup.cfg]
 
 [bumpversion:file:setup.py]
