+++ CONTRIBUTING.rst
@@ -100,7 +100,7 @@ Before you submit a pull request, check that it meets these guidelines:
 2. If the pull request adds functionality, the docs should be updated. Put
    your new functionality into a function with a docstring, and add the
    feature to the list in README.rst.
-3. The pull request should work for Python 3.5, 3.6, 3.7 and 3.8, and for PyPy. Check
+3. The pull request should work for Python 3.6, 3.7 and 3.8, and for PyPy. Check
    https://travis-ci.com/natemarks/treecrawl/pull_requests
    and make sure that the tests pass for all supported Python versions.
 
@@ -122,5 +122,8 @@ Then run::
 $ bump2version patch # possible: major / minor / patch
 $ git push
 $ git push --tags
+$ python setup.py sdist bdist_wheel
+$ twine upload --repository testpypi dist/*
+
 
 Travis will then deploy to PyPI if tests pass.
