+++ tests/testdata/test_upperificator/cities/expected/cold_cites.txt
@@ -1,0 +1,2 @@
+BOSTON
+ROCHESTER
