+++ tests/testdata/test_upperificator/cities/expected/warm_cities.txt
@@ -1,0 +1,2 @@
+AUSTIN
+TEMPE
