+++ tests/testdata/test_upperificator/cities/initial/cold_cites.txt
@@ -1,0 +1,2 @@
+boston
+rochester
