+++ tests/testdata/test_upperificator/cities/initial/warm_cities.txt
@@ -1,0 +1,2 @@
+austin
+tempe
