+++ tests/testdata/test_upperificator/pets/expected/janes_pets.txt
@@ -1,0 +1,2 @@
+LULU
+FIDO
