+++ tests/testdata/test_upperificator/pets/expected/johns_pets.txt
@@ -1,0 +1,2 @@
+BEN
+GERRY
