+++ tests/testdata/test_upperificator/pets/initial/janes_pets.txt
@@ -1,0 +1,2 @@
+lulu
+fido
