+++ tests/testdata/test_upperificator/pets/initial/johns_pets.txt
@@ -1,0 +1,2 @@
+ben
+gerry
